TEACHERUPDATES
1. shows enrolled courses

2. for each course it shows the sessions (Lectures) with the option to upload photos to each session.

sessions(Lectures) photos are stored in the database in a bucket storage.

3. CourseParticipants is functional.

* attendance is as follows:

x/y meaning that y is the number of overall sessions(lectures)
and x is the number of attended sessions for a studnet.
for example 1/3 means that there was 3 lectures and the student attended 1.

Next step is face recognition :)